requirements.txt
altgraph==0.17.4
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pywin32-ctypes==0.2.2
tk==0.1.0
